loader: i386 comparison between signed and unsigned integer expressions
    ../../common/multiboot2.c:106:16: error: comparison between signed and unsigned integ
er expressions [-Werror=sign-compare]
      for (i = 0; i < (rtag->mbh_size - sizeof (*rtag)) /
                    ^
../common/cons.c: In function 'getstr':
../common/cons.c:188:20: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (s - cmdstr < cmdstrsize - 1)
                    ^
zfsboot.c:319:34: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (read(fd, p, hdr.ex.a_text) != hdr.ex.a_text) {
                                  ^~
zfsboot.c:324:34: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (read(fd, p, hdr.ex.a_data) != hdr.ex.a_data) {
                                  ^~
zfsboot.c:333:35: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
    if (read(fd, p, hdr.ex.a_syms) != hdr.ex.a_syms) {
                                   ^~
zfsboot.c:345:23: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
    if (read(fd, p, x) != x) {
                       ^~
zfsboot.c:365:36: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
    if (read(fd, p, ep[i].p_filesz) != ep[i].p_filesz) {
                                    ^~
zfsboot.c:385:36: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (read(fd, p, es[i].sh_size) !=
                                    ^~
zfsboot.c: In function 'parttblread':
zfsboot.c:618:31: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (read(ppa->fd, buf, size) == size)
                               ^~
zfsboot.c: In function 'probe_partition':
zfsboot.c:639:10: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (len > sizeof (devname))
          ^
multiboot.c: In function 'multiboot_loadfile':
multiboot.c:132:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < (search_size / sizeof(uint32_t)); i++) {
                ^
linux.c: In function 'linux_loadkernel':
linux.c:160:28: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (linux_data_real_addr == -1) {
                            ^~
linux.c:221:40: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (archsw.arch_readin(fd, mem, text) != text) {
                                        ^~
biosdisk.c: In function 'bd_realstrategy':
biosdisk.c:1025:11: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (dblk < d_offset || dblk >= d_offset + disk_blocks)
           ^
biosdisk.c:1025:30: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (dblk < d_offset || dblk >= d_offset + disk_blocks)
                              ^~
biosdisk.c:1031:18: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (dblk + blks >= d_offset + disk_blocks) {
                  ^~
cc1: all warnings being treated as errors

+++ usr/src/boot/sys/boot/i386/common/cons.c
@@ -185,7 +185,7 @@ getstr(char *cmdstr, size_t cmdstrsize)
 			return;
 		default:
 			if (c >= 0x20 && c <= 0x7e) {
-				if (s - cmdstr < cmdstrsize - 1)
+				if ((size_t)(s - cmdstr) < cmdstrsize - 1)
 					*s++ = c;
 				putchar(c);
 			}
